need another go-around with int conversions
sigh

+++ bld/cdc32/qmsgsql.cpp
@@ -1520,7 +1520,7 @@ sql_has_msg_recently(vc uid, long num_seconds)
     res = sql_simple("select max(date) from gi where assoc_uid = ?1 and is_sent isnull", huid);
     if(res.num_elems() != 1)
         return false;
-    int64_t dm = res[0][0];
+    int64_t dm = (long)res[0][0];
     if(time(0) - dm < num_seconds)
         return true;
     return false;
